+++ AvgScript/src/extension.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 
 import { colorProvider } from './functions/color';
-import { assetsListPanel, commandAppendDialogue, commandAppendDialogue_impl, commandBasePath, commandBasePath_impl, commandDeleteDub, commandDeleteDub_impl, commandGetAssetsList, commandGetAssetsList_impl, commandGetDubList, commandGetDubList_impl, commandRefreshAssets, commandRefreshAssets_impl, commandReplaceScript, commandReplaceScript_impl, commandShowDialogueFormatHint, commandShowDialogueFormatHint_impl, commandShowHibiscusDocument, commandShowHibiscusDocument_impl, commandShowJumpFlow, commandShowJumpFlow_impl, commandUpdateCommandExtension, commandUpdateCommandExtension_impl, commandUpdateDub, commandUpdateDub_impl, dubListPanel, formatHintPanel, jumpFlowPanel } from './functions/command';
+import { assetsListPanel, commandAppendDialogue, commandAppendDialogue_impl, commandBasePath, commandBasePath_impl, commandDeleteDub, commandDeleteDub_impl, commandGetAssetsList, commandGetAssetsList_impl, commandGetDubList, commandGetDubList_impl, commandPasteDub, commandPasteDub_impl, commandRefreshAssets, commandRefreshAssets_impl, commandReplaceScript, commandReplaceScript_impl, commandShowDialogueFormatHint, commandShowDialogueFormatHint_impl, commandShowHibiscusDocument, commandShowHibiscusDocument_impl, commandShowJumpFlow, commandShowJumpFlow_impl, commandUpdateCommandExtension, commandUpdateCommandExtension_impl, commandUpdateDub, commandUpdateDub_impl, dubListPanel, formatHintPanel, jumpFlowPanel } from './functions/command';
 import { atCommands, fileName, fileSuffix, langPrefix, required, settingsParam, sharpCommands } from './functions/completion';
 import { debuggerFactory, debuggerProvider, debuggerTracker } from './functions/debugger';
 import { diagnosticThrottle, diagnosticsCollection } from './functions/diagnostic';
@@ -54,6 +54,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
 	vscode.commands.registerCommand(commandUpdateDub, commandUpdateDub_impl);
 	vscode.commands.registerCommand(commandDeleteDub, commandDeleteDub_impl);
+	vscode.commands.registerCommand(commandPasteDub, commandPasteDub_impl);
 
 	//--------------------
 	// Init Command